Refactor some IO components; move/rename some bits.
Replace `BPETokenVocabTrainer` with `BinaryPairVocabTrainer` and refactor Tiktoken vocab I/O to `WordMapTokenVocab` for improved modularity and consistency.
