+++ crates/bpetok/src/training/mod.rs
@@ -4,3 +4,5 @@ pub mod pair_index;
 pub mod trainer;
 pub mod word;
 pub mod word_count;
+
+pub use trainer::*;
