+++ crates/bpetok/src/training/trainer.rs
@@ -14,7 +14,7 @@ use std::cmp::Ordering;
 
 /// A training for [`Tokenizer`]s.
 #[derive(Debug)]
-pub struct BPETokenVocabTrainer {
+pub struct BinaryPairVocabTrainer {
     /// The regex pattern used for text splitting.
     pub pattern: String,
 
@@ -25,8 +25,8 @@ pub struct BPETokenVocabTrainer {
     pub parallel: bool,
 }
 
-impl BPETokenVocabTrainer {
-    /// Creates a new [`BPETokenVocabTrainer`].
+impl BinaryPairVocabTrainer {
+    /// Creates a new [`BinaryPairVocabTrainer`].
     ///
     /// # Arguments
     /// * `vocab_size` - The desired vocabulary size; must be >= 256 (the size of the u8 space).
@@ -49,7 +49,7 @@ pub struct TrainResults<T: TokenType> {
     pub pair_vocab: PairMapTokenVocab<T>,
 }
 
-impl BPETokenVocabTrainer {
+impl BinaryPairVocabTrainer {
     /// Sets the vocab size.
     ///
     /// # Arguments
@@ -332,7 +332,7 @@ mod tests {
     use crate::decoders::token_decoder::TokenDecoder;
     use crate::encoders::token_encoder::TokenEncoder;
     use crate::encoders::unified_encoder::UnifiedVocabEncoder;
-    use crate::training::trainer::{BPETokenVocabTrainer, MergeJob, TrainResults};
+    use crate::training::trainer::{BinaryPairVocabTrainer, MergeJob, TrainResults};
     use crate::types::{check_is_send, check_is_sync};
     use crate::vocab::TokenVocabIndex;
     use crate::vocab::unified_vocab::UnifiedTokenVocab;
@@ -343,7 +343,7 @@ mod tests {
 
     #[test]
     fn test_tokenizer_options() {
-        let options = BPETokenVocabTrainer::new_with_vocab_size(1000);
+        let options = BinaryPairVocabTrainer::new_with_vocab_size(1000);
         assert_eq!(options.vocab_size, 1000);
         assert_eq!(options.pattern, DEFAULT_PATTERN);
         assert_eq!(options.parallel, DEFAULT_PARALLEL);
@@ -361,7 +361,7 @@ mod tests {
     #[test]
     #[should_panic(expected = "regex pattern compilation failed")]
     fn test_tokenizer_options_bad_pattern() {
-        let _ = BPETokenVocabTrainer::new_with_vocab_size(1000).with_pattern(r"(");
+        let _ = BinaryPairVocabTrainer::new_with_vocab_size(1000).with_pattern(r"(");
     }
 
     #[test]
@@ -380,7 +380,7 @@ mod tests {
         type C = u32;
         type K = CompactString;
 
-        let options = BPETokenVocabTrainer::new_with_vocab_size(1000).with_parallel(parallel);
+        let options = BinaryPairVocabTrainer::new_with_vocab_size(1000).with_parallel(parallel);
 
         let samples = vec![
             "hello world",
