+++ crates/bpetok/src/vocab/mod.rs
@@ -1,9 +1,10 @@
 //! # Vocabulary
 
 pub mod pair_vocab;
-pub mod tiktoken_io;
 pub mod unified_vocab;
 pub mod vocab_index;
 pub mod word_vocab;
 
+pub use unified_vocab::*;
 pub use vocab_index::*;
+pub use word_vocab::*;
